% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ems_duckdb.R
\name{attach_historic_data}
\alias{attach_historic_data}
\title{Load the historic ems database as a tbl}
\usage{
attach_historic_data(con = NULL)
}
\value{
A dplyr connection to the duckdb database. See \code{\link[dplyr]{tbl}} for more.
}
\description{
You can then use dplyr verbs such as \code{\link[dplyr]{filter}},
\code{\link[dplyr]{select}}, \code{\link[dplyr]{summarize}}, etc. For basic
importing of historic data based on \code{ems_id}, \code{date}, and \code{parameter},
you can use the function \code{\link{read_historic_data}}
}
\examples{
\dontrun{
library(dplyr)

con <- connect_historic_db()

hist_tbl <- attach_historic_data(con)
result <- hist_tbl \%>\%
 group_by(EMS_ID) \%>\%
 summarise(max_date = max(COLLECTION_START))
collect(result)

disconnect_historic_db(con)
}
}
