% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ops.R
\name{bind_ems_data}
\alias{bind_ems_data}
\title{Combine rows from different EMS data sets}
\usage{
bind_ems_data(...)
}
\arguments{
\item{...}{the datasets you want to combine}
}
\value{
a data frame with rows of inputs combined
}
\description{
Mostly used to combine '2yr' and 'historic' data obtained through
\link{get_ems_data} and \link{read_historic_data}.
}

