% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onLoad.R
\name{burn_it_down}
\alias{burn_it_down}
\title{Destroy data cache}
\usage{
burn_it_down()
}
\value{
TRUE
}
\description{
Use this if you are getting odd errors when trying to get data or update dates
from the cache
}

