% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{connect}
\alias{connect}
\title{Connect to EMS and get the Auth token}
\usage{
connect(usr, pwd, proxies = NULL, server = "old", server_url = NULL)
}
\arguments{
\item{usr}{String, FOQA username}

\item{pwd}{String, FOQA password}

\item{proxies}{List containing the system proxy information. The list contains attributes "url", "port", "usr", "pwd"}

\item{server}{String, temporary parameter that let you choose an API server. Currently "old" (2-node), "new" (3-node now but eventually grow to 24 nodes)}
}
\value{
a Connection object.
}
\description{
Connect to EMS and get the Auth token
}

