% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ems_sqlite.R
\name{connect_historic_db}
\alias{connect_historic_db}
\title{Create a database connection to the historic database}
\usage{
connect_historic_db(db_path = NULL)
}
\arguments{
\item{db_path}{path to the historic SQLite database. In most cases
it does not need to be specified as it uses the default location
set by \code{rems}.}
}
\value{
a DBIConnection object for communicating with the SQLite database
}
\description{
This creates a DBI connection to the SQLite database
that holds the historic data table. You should call
\code{\link[=disconnect_historic_db]{disconnect_historic_db()}} when you are finished
querying the database.
}
\details{
See \code{\link[=attach_historic_data]{attach_historic_data()}} for examples.
}
\seealso{
\code{\link[DBI:dbConnect]{DBI::dbConnect()}}
}
