% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dictionaries.R
\docType{data}
\name{data_dictionaries}
\alias{data_dictionaries}
\alias{ems_coll_methods}
\alias{ems_parameters}
\alias{ems_sample_classes}
\alias{ems_location_samples}
\alias{ems_species}
\alias{ems_units}
\title{Data Dictionaries for EMS data}
\format{
Objects of class \code{tbl_df} (inherits from \code{tbl, data.frame}).

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 28 rows and 2 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6926 rows and 7 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 16 rows and 2 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 332 rows and 6 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2917 rows and 3 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 231 rows and 3 columns.
}
\source{
\href{https://catalogue.data.gov.bc.ca/dataset/bc-environmental-monitoring-system-results#additional-information}{BC Data Catalogue: EMS Results}
}
\usage{
ems_coll_methods

ems_parameters

ems_sample_classes

ems_location_samples

ems_species

ems_units
}
\description{
Various data dictionaries (lookup tables)
}
\keyword{datasets}
