% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ems_duckdb.R
\name{disconnect_historic_db}
\alias{disconnect_historic_db}
\title{Close the connection to the historic database}
\usage{
disconnect_historic_db(con)
}
\arguments{
\item{con}{DBI connection object, most likely created by \code{\link[=connect_historic_db]{connect_historic_db()}}.}
}
\value{
\code{TRUE}, invisibly
}
\description{
Close the connection to the historic database
}
