% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flight.R
\docType{data}
\name{.default_exclude_dirs}
\alias{.default_exclude_dirs}
\title{Default directories to exclude from tree traversal}
\format{
An object of class \code{character} of length 11.
}
\usage{
.default_exclude_dirs
}
\description{
These directories are excluded from tree traversal by default,
but can be overridden by setting options.
}
\keyword{internal}
