% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ems_sqlite.R
\name{download_historic_data}
\alias{download_historic_data}
\title{Download and store the large historic ems data}
\usage{
download_historic_data(n = 1e+06, force = FALSE)
}
\arguments{
\item{n}{the chunk size to use to iteratively read and store the historic data}
}
\value{
TRUE
}
\description{
Download and store the large historic ems data
}

