% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ems_locations.R
\name{ems_location_features}
\alias{ems_location_features}
\title{Get the data field names and types for the EMS monitoring locations}
\usage{
ems_location_features()
}
\value{
A data.frame of column names and their properties
}
\description{
This can help you know what to query in [ems_locations()]
}
