% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ems_locations.R
\name{ems_locations}
\alias{ems_locations}
\title{Get EMS monitoring locations}
\usage{
ems_locations(epsg = 3005, query = "LOCATION_TYPE_CD='21'",
  sf = TRUE)
}
\arguments{
\item{epsg}{The coordinate reference system}

\item{query}{Query string in ECQL syntax. Default is to only get monitoring
location types of 'River Stream or Creek' (`LOCATION_TYPE_CD='21'`). See the
[geoserver documentation](http://docs.geoserver.org/stable/en/user/filter/ecql_reference.html#filter-ecql-reference)}

\item{sf}{should an `sf` spatial object be returned (default `TRUE`),
or just a `data.frame`?}
}
\value{
an `sf` object if `sf = TRUE`, otherwise a data.frame of the attributes
}
\description{
Get an `sf` object of EMS monitoring locations
}
\examples{
# Get river stream or creek locations with a purpose code of 2 (TREND):
ems_locations(query = "LOCATION_TYPE_CD='21' AND LOCATION_PURPOSE_CD='2'")
}
