% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ems_posix_numeric}
\alias{ems_posix_numeric}
\title{Convert an integer representing a Unix date/time to POSIXct (R date/time) class}
\usage{
ems_posix_numeric(x)
}
\arguments{
\item{x}{Datetime integer}
}
\value{
POSIXct vector
}
\description{
Convert an integer representing a Unix date/time to POSIXct (R date/time) class
}
