% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flight.R
\name{escape_pattern}
\alias{escape_pattern}
\title{Escape special characters in a string for regular expression matching}
\usage{
escape_pattern(pattern)
}
\arguments{
\item{pattern}{Character string to escape}
}
\value{
Character string with special characters escaped
}
\description{
Escapes special characters in a string to allow safe use in regular expressions.
This is a more robust replacement for the custom treat_spchar function.
}
\examples{
escape_pattern("file.txt")  # escapes the dot
escape_pattern("data[2020]") # escapes the brackets

}
