% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ops.R
\name{filter_ems_data}
\alias{filter_ems_data}
\title{Simple filtering of ems data by emsid and or dates.}
\usage{
filter_ems_data(x, emsid = NULL, from_date = NULL, to_date = NULL)
}
\arguments{
\item{x}{The ems dat frame to filter}

\item{emsid}{A character string of the ems id of interest}

\item{from_date}{A date string in a standard unambiguos format (e.g., "YYYY/MM/DD")}

\item{to_date}{A date string in a standard unambiguos format (e.g., "YYYY/MM/DD")}
}
\value{
A filtered data frame
}
\description{
Simple filtering of ems data by emsid and or dates.
}

