% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flight.R
\name{fl_request}
\alias{fl_request}
\title{Request field information from the API}
\usage{
fl_request(flt, parent)
}
\arguments{
\item{flt}{A Flight object}

\item{parent}{A list with field group information}
}
\value{
A list containing field and field group information
}
\description{
Requests field information from the API for a given parent node.
}
