% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flt_query.R
\name{flt_query}
\alias{flt_query}
\title{Create Flight Query}
\usage{
flt_query(conn, ems_name, data_file = NULL)
}
\description{
Create a flight query object for use in \code{run} function
}
\details{
This function uses a connection object to create a flight query object.
These can be manipulated using the \code{filter}, \code{select} and other
functions before passing to \code{run} to execute and return a dataframe.
This object won't work by itself - you will need to set a database, select elements
etc for it to work.

 @param conn A connection object generated by \code{connect}
 @param ems_name A string, indicating the name of the EMS server
}
