% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FP_river_sites.R
\name{FP_river_sites}
\alias{FP_river_sites}
\title{Canada-BC Long-term river monitoring sites}
\usage{
FP_river_sites()
}
\value{
a character vector of ems ids
}
\description{
Get the EMS_IDs of all of the Canada-BC long-term river monitoring sites.
}
