% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_preset_fieldtree.R
\name{generate_preset_fieldtree}
\alias{generate_preset_fieldtree}
\title{Generate Preset Fieldtree}
\usage{
generate_preset_fieldtree(qry)
}
\arguments{
\item{qry}{A fltquery object generated by \code{flt_query}}
}
\value{
A FltQuery object updated with fieldtree paths for use downstream.
}
\description{
Initialize a \code{fltquery} object with available tree objects.
}
\details{
This function will take the existing tree available in the object and
attempt to generate a fieldtree based on the available root paths in the object.
Typically this woule be the first call after initializing a \code{fltquery} object,
as calls to \code{update_fieldtree} or \code{select} may not even work until a
minimal fieldtree is available.
}
\examples{
\dontrun{
conn <- connect("joe.bloggs", "mypassword")
qry <- flt_query(conn, "my-ems", data_file = "metadata.db")
qry <- generate_preset_fieldtree(qry)
}

}
