% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{get_cache_date}
\alias{get_cache_date}
\title{Get the date(s) when ems data was last updated locally.}
\usage{
get_cache_date(which)
}
\arguments{
\item{which}{The data for which you want to check it's cache date. "2yr", "4yr", or "historic}
}
\value{
The date the data was last updated (if it exists in your cache)
}
\description{
Get the date(s) when ems data was last updated locally.
}

