% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ems_data.R
\name{get_ems_data}
\alias{get_ems_data}
\title{get EMS data from BC Data Catalogue}
\usage{
get_ems_data(which = "current")
}
\arguments{
\item{which}{Do you want current (past 2 years) or historic data?}
}
\value{
a data frame
}
\description{
get EMS data from BC Data Catalogue
}

