% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-shinyrems.R
\name{get_ems_lookup}
\alias{get_ems_lookup}
\title{get EMS data lookup table}
\usage{
get_ems_lookup(which = "2yr", ask = TRUE)
}
\arguments{
\item{which}{Defaults to \code{"2yr"} (past 2 years). You can also specify "4yr"
to get the past four years of data.}

\item{ask}{should the function ask for your permission to cache data on your computer?
Default \code{TRUE}}
}
\value{
a data frame
}
\description{
This function generates and caches a lookup table of the chosen data ('2yr' or '4yr') and imports it into your R session.
If the cached data are more current than than the lookup cache, you will be prompted to update it.
}
\details{
this function is only used when running the ShinyRems app.

\code{"EMS_ID", "MONITORING_LOCATION", "PERMIT", "PARAMETER_CODE", "PARAMETER",
"LONGITUDE" "LATITUDE", "FROM_DATE", "TO_DATE"}
}
\keyword{internal}
