% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flight.R
\name{get_exclude_dirs}
\alias{get_exclude_dirs}
\title{Get directories to exclude from tree traversal}
\usage{
get_exclude_dirs()
}
\value{
Character vector of directory names to exclude
}
\description{
Returns the list of directories to exclude from tree traversal.
This can be customized by setting the 'flightems.exclude_dirs' option.
}
\examples{
# Get current exclusion list
get_exclude_dirs()

# Set custom exclusion list
options(flightems.exclude_dirs = c("Profiles", "Weather Information"))
get_exclude_dirs()

# Reset to defaults
options(flightems.exclude_dirs = NULL)
get_exclude_dirs()

}
