% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_top_fltquery.R
\name{get_top}
\alias{get_top}
\title{Restrict to Top N Results}
\usage{
get_top(qry, ...)
}
\description{
Alters a \code{fltquery} object to only retrieve the first n rows
}
\details{
This function alters a \code{fltquery} object so that it will only return
the first n records, as defined by the filtering criteria applied by \code{filter}

 @param qry A fltquery object generated by \code{flt_query}
 @param n A numeric, indicating the number of rows requested
 @return A FltQuery object that will only return n rows when run.
}
\examples{
\dontrun{
conn <- connect("joe.bloggs", "mypassword")
qry <- flt_query(conn, "my-ems", data_file = "metadata.db")
qry <- generate_preset_fieldtree(qry)
}

}
