% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ems_data.R
\name{get_update_date}
\alias{get_update_date}
\title{Get the date(s) when ems data was last updated.}
\usage{
get_update_date(which)
}
\arguments{
\item{which}{The data for which you want to check it's update date. "current" or "historic}
}
\value{
The date the data was last updated (if it exists in your cache)
}
\description{
Get the date(s) when ems data was last updated.
}

