% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lt_lake_ids.R
\docType{data}
\name{lt_lake_ids}
\alias{lt_lake_ids}
\title{Active lake monitoring sites as of 09-03-2021.}
\format{
A data frame with 1307 rows and 3 variables
\describe{
\item{EMS_ID}{EMS IDs of sites}
\item{MONITORING_LOCATION}{names of sites}
\item{REQUISTION_ID}{Requisition IDS of sites}
}
}
\source{
internal data source
}
\usage{
lt_lake_ids
}
\description{
A dataset containing the EMS ID, Site Name, and Requisition IDs of all active lake monitoring stations.
}
\keyword{datasets}
