% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{lt_lake_req}
\alias{lt_lake_req}
\title{Long-term lake monitoring sites}
\usage{
lt_lake_req(data = NULL)
}
\arguments{
\item{data}{loads a site ID lookup table}
}
\value{
a character vector of requisitions ids
}
\description{
Get the REQ_IDs of all of the long-term lake monitoring sites.
}
