% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ems_sqlite.R
\name{read_historic_data}
\alias{read_historic_data}
\title{Read historic ems data into R.}
\usage{
read_historic_data(emsid = NULL, parameter = NULL, from_date = NULL,
  to_date = NULL, cols = "wq")
}
\arguments{
\item{emsid}{the EMS_ID(s) of the stations you want}

\item{parameter}{the parameter(s) of the stations you want}

\item{from_date}{start date}

\item{to_date}{end date}

\item{cols}{colums. Default "wq". See \code{link{get_ems_data}} for further documentation.}
}
\value{
a data frame of the results
}
\description{
You will need to have run \code{\link{download_historic_data}} before you
can use this function
}

