% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{reconnect}
\alias{reconnect}
\title{Refresh Connection}
\usage{
reconnect(conn)
}
\arguments{
\item{conn}{A connection object created by the \code{connect} function.}
}
\value{
A refreshed connection object like the one returned by \code{connect}
}
\description{
Refresh the connection object in case auth token expires.
}
\details{
This function will re-use the stored connection credentials in the connection object to regenerate the authentication token.
}
\examples{
\dontrun{
con <- connect("joe.bloggs", "mypassword")
sys.Sleep(3600)
con <- reconnect(con)
}

}
