% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{remove_data_cache}
\alias{remove_data_cache}
\title{Remove cached EMS data from your computer}
\usage{
remove_data_cache(which)
}
\arguments{
\item{which}{which data to remove? Either \code{"2yr"}, \code{"4yr"},
\code{"historic"}, or \code{"all"}.}
}
\description{
Remove cached EMS data from your computer
}
