% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rems-package.R
\docType{package}
\name{rems-package}
\alias{rems}
\alias{rems-package}
\title{rems: Get Data from British Columbia's Environmental Monitoring
    System}
\description{
Download and parse data from B.C.'s Environmental
    Monitoring System.
}
\author{
\strong{Maintainer}: Andy Teucher \email{andy.teucher@gov.bc.ca}

Other contributors:
\itemize{
  \item Province of British Columbia [copyright holder]
}

}
\keyword{internal}
