% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{request}
\alias{request}
\title{Arbitrary Request}
\usage{
request(
  conn,
  rtype = c("GET", "POST", "DELETE"),
  uri_keys = NULL,
  uri_args = NULL,
  headers = NULL,
  body = NULL,
  jsondata = NULL
)
}
\arguments{
\item{conn}{A connection object generated by \code{connect}}

\item{rtype}{A string indicating the kind of HTTP request to be performed - accepts GET, POST and DELETE}

\item{uri_keys}{A two-element string vector indicating which URI to use - see common.R}

\item{uri_args}{A list of arguments to be passed to the URI}

\item{headers}{A list of headers to be used by the GET/POST/DELETE request, per the \code{.headers} argument}

\item{body}{A string, indicating the body to be passed to the request. Do not pass this and jsondata simultaneously.}

\item{jsondata}{A string, indicating the JSON body to be passed. See above.}
}
\value{
Returns the raw GET/POST/DELETE output from the request.
}
\description{
Perform an arbitrary GET, POST or DELETE function using a connection  object.
}
\details{
For advanced users only, this function allows you to perform an arbitrary GET, POST or DELETE
request on the low-level API. Allows for full configurability of the request, including
setting the keys, arguments, headers, and JSON payload.
}
