% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_ems_instances.R
\name{retrieve_ems_instances}
\alias{retrieve_ems_instances}
\title{Retrieve the names of EMS Instances}
\usage{
retrieve_ems_instances(conn)
}
\arguments{
\item{conn}{A connection object generated by \code{connect} with an active auth token}
}
\value{
a string vector of the available EMS names for your user
}
\description{
Convenience function to check what EMS Names are available to your user
}
\details{
The flt_query object requires a EMS name, but the ones available to you may not be immediately
obvious - this function retrieves those names.
}
