% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{save_ems_data}
\alias{save_ems_data}
\title{Save EMS data as a csv file}
\usage{
save_ems_data(obj = NULL, which = NULL, filename = NULL, ...)
}
\arguments{
\item{obj}{The name of an object in your environment}

\item{which}{"2yr", "4yr", or "historic"}

\item{filename}{the name of the file you are writing to}

\item{...}{other options passed on to \code{read_csv}}
}
\value{
the object, invisibly
}
\description{
You must specify either an object in your environment (via \code{obj}),
or one of \code{"2yr"}, \code{"4yr"}, or \code{"historic"} (via \code{which}),
but not both.
}

