% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytic.R
\name{search_param}
\alias{search_param}
\title{Search Global Parameters}
\usage{
search_param(
  anal,
  keyword,
  category = c("full", "physical", "logical"),
  flight_record = NULL,
  group = NULL,
  max_results = NULL
)
}
\arguments{
\item{anal}{An analytic object (an element of list returned by tseries_query)}

\item{keyword}{A character string to search for in the parameter names.}

\item{category}{Kind of parameter to search. Defaults to 'full' (all available.)}

\item{flight_record}{Optional flight record to restrict search to.}

\item{group}{A group ID to restrict the search to. Defaults to all group IDs.}

\item{max_results}{Defaults to 200 - if set to 0, it will return all results.}
}
\value{
A list of parameters lists (list of lists) with the following fields:
\itemize{
\item id: EMS GUID of the parameter
\item name: Name of the parameter
\item description: Description of the parameter
\item units: the units of the parameter, if available
\item metadata: any associated metadata text, if available
\item ems_id: EMS ID of the analytic
}
}
\description{
Search for global parameters matching a substring.
}
\details{
This function uses the analytic element of a time series query
to search for global parameters matching a substring. It will return
a list containing all the parameters that match the substring, as well
as their EMS GUIDs, names, descriptions, and units.
}
\examples{
\dontrun{
# Connect to EMS
conn <- connect("http://localhost:8080")
# Get a time series query object
qry <- tseries_query(conn, 1)
# Search for parameters with keyword "weather"
prm <- search_param(qry$analytic, "weather")
}


}
