% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flt_query.R
\name{set_database.FltQuery}
\alias{set_database.FltQuery}
\title{Set database}
\usage{
\method{set_database}{FltQuery}(qry, name)
}
\description{
Set the database for the flight query object.
}
\details{
This
}
