% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ops.R
\name{set_ems_tz}
\alias{set_ems_tz}
\title{Set a date-time to the timezone used by EMS}
\usage{
set_ems_tz(x)
}
\arguments{
\item{x}{a date-time-like object or character string
in a standard unambiguous format (e.g., \code{"YYYY/MM/DD"}).}
}
\value{
\code{POSIXct} in the updated timezone.
}
\description{
Sets the timezone to PST (UTC -8 hrs; Etc/GMT+8), which is what is used
in EMS. Does not change the clock time of the input, rather
forces the timezone to be UTC -8 hrs.
}
\seealso{
\code{\link[lubridate:force_tz]{lubridate::force_tz()}}, \code{\link[=OlsonNames]{OlsonNames()}}
}
