% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{standardize_mdl_units}
\alias{standardize_mdl_units}
\title{Standardize MDL}
\usage{
standardize_mdl_units(data)
}
\arguments{
\item{data}{an ems data frame containing at least the
columns \verb{"UNIT", "METHOD_DETECTION_LIMIT", "MDL_UNIT"}}
}
\value{
data frame with MDLs standardized to UNITs (where possible)
}
\description{
Standardize MDL
}
