% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_query.R
\name{tseries_query}
\alias{tseries_query}
\title{Time Series Query}
\usage{
tseries_query(conn, ems_name, data_file = NULL, flight_record_searched = NULL)
}
\arguments{
\item{conn}{A connect object created by the connect function.}

\item{ems_name}{A string, the name of the EMS instance being connected to.}

\item{data_file}{A string, the name of the db file to store metadata to.}

\item{flight_record_searched}{optional numeric flight record to search params}
}
\description{
Instantiate a time series query object
}
\details{
This function creates a time series query object that can be used to query
time series (i.e. parameter level) data from EMS. If you need physical
parameter data, you will need to set a specific flight record to search
to return those parameters.
}
