% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ems_duckdb.R
\name{attach_historic_data}
\alias{attach_historic_data}
\title{Load the historic ems database as a tbl}
\usage{
attach_historic_data()
}
\value{
A dplyr connection to the duckdb database. See \code{\link[dplyr]{src_sql}} for more.
}
\description{
You can then use dplyr verbs such as \code{\link[dplyr]{filter}},
\code{\link[dplyr]{select}}, \code{\link[dplyr]{summarize}}, etc. For basic
importing of historic data based on \code{ems_id}, \code{date}, and \code{parameter},
you can use the function \code{\link{read_historic_data}}
}
\examples{
\dontrun{
library(dplyr)
foo <- attach_historic_data()
bar <- foo \%>\%
  group_by(EMS_ID) \%>\%
  summarise(max_date = max(COLLECTION_START))
baz <- collect(bar)
baz$max_date <- as.POSIXct(baz$max_date, origin = "1970/01/01", tz = "Etc/GMT+8")
}
}
