% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{connect}
\alias{connect}
\title{Create EMS Connection Object}
\usage{
connect(
  usr,
  pwd,
  proxies = NULL,
  server = c("prod", "cluster", "stable", "beta", "nightly", "ctc"),
  server_url = NULL
)
}
\arguments{
\item{usr}{A string, FOQA username (like firstname.lastname)}

\item{pwd}{A string, FOQA password}

\item{proxies}{A list containing the system proxy information, with attributes "url", "port", "usr", "pwd"}

\item{server}{A string. Lets you choose an API server. Should be one of 'prod', 'cluster', 'stable', 'beta', 'nightly', or 'ctc'.}

\item{server_url}{An optional string. If this is passed, it will be used instead of the internal server definitions matched by the 'server' argument.}
}
\value{
a Connection object to be used in FltQuery.
}
\description{
\code{connect} Creates a connection object to an EMS server for use in FltQuery objects.
}
\details{
This function creates a connection object, including an authentication token.
These can be passed into a \code{fltquery} or \code{tsquery} function to start the process of retrieving data.
These functions take unmasked usernames and passwords - it is recommended that you either store your credentials
securely, or wrap this function in something like \code{rstudioapi::askForPassword()} to avoid disclosing
your credentials.
}
\examples{
# connect in a standard way, using no proxy and connecting to prod by default.
\dontrun{
con <- connect("joe.bloggs", "mypassword")
qry <- flt_query(conn = con, ems_name = "my-ems", data_file = "file.db")
}
# connect using the beta build, and a proxy
\dontrun{
con <- connect("joe.bloggs", "mypassword",
              proxies = list(url = "http://myproxy.com", port = 8080, usr = "joebloggs", pwd = "mypassword"),
              server = 'beta')
}

}
