% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb_create.R
\name{create_rems_duckdb}
\alias{create_rems_duckdb}
\title{Create historic table in duckdb}
\usage{
create_rems_duckdb(csv_file, db_path)
}
\arguments{
\item{csv_file}{path to csv file}

\item{db_path}{where to create the duckdb}
}
\value{
\code{TRUE} (invisibly)
}
\description{
Create historic table in duckdb
}
