% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ops.R
\name{filter_ems_data}
\alias{filter_ems_data}
\title{Simple filtering of ems data by emsid and or dates.}
\usage{
filter_ems_data(
  x,
  emsid = NULL,
  parameter = NULL,
  param_code = NULL,
  from_date = NULL,
  to_date = NULL,
  req_id = NULL
)
}
\arguments{
\item{x}{The ems data frame to filter}

\item{emsid}{A character vector of the ems id(s) of interest}

\item{parameter}{a character vector of parameter names}

\item{param_code}{a character vector of parameter codes}

\item{from_date}{A \code{Date}, \code{POSIXct}, \code{POSIXlt}, or a \code{character} string in a standard unambiguous format (e.g., \code{"YYYY/MM/DD"})}

\item{to_date}{A \code{Date}, \code{POSIXct}, \code{POSIXlt}, or a \code{character} string in a standard unambiguous format (e.g., \code{"YYYY/MM/DD"})}

\item{req_id}{A character vector of requisition id(s) of interest. See \code{\link[=lt_lake_req]{lt_lake_req()}} for a list of long-term lakes monitoring sites REQUISITION IDs}
}
\value{
A filtered data frame
}
\description{
Simple filtering of ems data by emsid and or dates.
}
