% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flt_query.R
\name{flt_query}
\alias{flt_query}
\title{Create Flight Query}
\usage{
flt_query(conn, ems_name, data_file = NULL)
}
\arguments{
\item{conn}{A connection object generated by \code{connect}}

\item{ems_name}{A string, indicating the name of the EMS server. Will return an informative error if you use the wrong name.}

\item{data_file}{An optional string, indicating the name of the .db file where the query tree metadata will be stored locally.}
}
\value{
A FltQuery object to be manipulated and then run.
}
\description{
Create a flight query object for use in \code{run} function
}
\details{
This function uses a connection object to create a flight query object.
These can be manipulated using the \code{filter}, \code{select} and other
functions before passing to \code{run} to execute and return a dataframe.
This object won't work by itself - you will need to set a database, select elements
etc for it to work.
}
\examples{
\dontrun{
conn <- connect("joe.bloggs", "mypassword")
qry <- flt_query(conn, "my-ems", data_file = "metadata.db")
}

}
