% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flt_query.R
\name{flt_query}
\alias{flt_query}
\title{Create Flight Query}
\usage{
flt_query(conn, ems_name, data_file = NULL)
}
\arguments{
\item{conn}{a connection object}

\item{ems_name}{a string indicating the EMS server being used}

\item{data_file}{a SQLite db file to be created to store the field and db trees.}
}
\value{
this function returns a connection object as a list
}
\description{
Create a flight query object
}
\details{
This function creates a flight query object that can be used to retrieve data.
}
