% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ems_data.R
\name{get_ems_data}
\alias{get_ems_data}
\title{get EMS data from BC Data Catalogue}
\usage{
get_ems_data(which = "current")
}
\arguments{
\item{which}{Do you want "current" (past 2 years; default) or "historic" data?
Currently only supports current as the historic files are really big and
need special handling that hasn't been implemented yet}
}
\value{
a data frame
}
\description{
get EMS data from BC Data Catalogue
}

