% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ems_data.R
\name{get_ems_data}
\alias{get_ems_data}
\title{get EMS data from BC Data Catalogue}
\usage{
get_ems_data(which = "2yr", n = Inf, cols = "wq", force = FALSE,
  ask = TRUE, dont_update = FALSE)
}
\arguments{
\item{which}{Defaults to \code{"2yr"} (past 2 years). You can also specify "4yr"
to get the past four years of data. If you want historic data, use the
\code{\link{download_historic_data}} and \code{\link{read_historic_data}} functions.}

\item{n}{how many rows of the data do you want to load? Defaults to all (\code{n = Inf}).}

\item{cols}{which subset of columns to read. Can be \code{"all"} which reads all
columns, \code{"wq"} (default) which returns a predefined subset of columns common
for water quality analysis, or a character vector of column names (see details below).}

\item{force}{Default \code{FALSE}. Setting to \code{TRUE} will download new data even
if it's not out of date on your computer.}

\item{ask}{should the function ask for your permission to cache data on your computer?
Default \code{TRUE}}

\item{dont_update}{should the function avoid updating the data even if there is a newer
version available? Default \code{FALSE}}
}
\value{
a data frame
}
\description{
EMS data are distributed through the \href{https://catalogue.data.gov.bc.ca/dataset/bc-environmental-monitoring-system-results}{BC Data Catalogue}
under the \href{http://www.data.gov.bc.ca/local/dbc/docs/license/OGL-vbc2.0.pdf}{Open Government License - British Columbia}.
This function downloads the chosen data ('historic' - 1964-2014, or '2yr' or '4yr') and imports it into your R session. It also caches the data so
subsequent loads are much faster - if the data in the Data Catalogue are more
current than that in your cache, you will be prompted to update it.
}
\details{
cols can specify any of the following column names as a character vector:

\code{"EMS_ID", "MONITORING_LOCATION", "LATITUDE", "LONGITUDE", "LOCATION_TYPE",
"COLLECTION_START", "COLLECTION_END", "LOCATION_PURPOSE", "PERMIT",
"PERMIT_RELATIONSHIP", "DISCHARGE_TO", "REQUISITION_ID", "SAMPLING_AGENCY",
"ANALYZING_AGENCY", "COLLECTION_METHOD", "SAMPLE_CLASS", "SAMPLE_STATE",
"SAMPLE_DESCRIPTOR", "PARAMETER_CODE", "PARAMETER", "ANALYTICAL_METHOD_CODE",
"ANALYTICAL_METHOD", "RESULT_LETTER", "RESULT", "UNIT", "METHOD_DETECTION_LIMIT",
"QA_INDEX_CODE", "UPPER_DEPTH", "LOWER_DEPTH", "TIDE", "AIR_FILTER_SIZE",
"AIR_FLOW_VOLUME", "FLOW_UNIT", "COMPOSITE_ITEMS", "CONTINUOUS_AVERAGE",
"CONTINUOUS_MAXIMUM", "CONTINUOUS_MINIMUM", "CONTINUOUS_UNIT_CODE",
"CONTINUOUS_DURATION", "CONTINUOUS_DURATION_UNIT", "CONTINUOUS_DATA_POINTS",
"TISSUE_TYPE", "SAMPLE_SPECIES", "SEX", "LIFE_STAGE", "BIO_SAMPLE_VOLUME",
"VOLUME_UNIT", "BIO_SAMPLE_AREA", "AREA_UNIT", "BIO_SIZE_FROM",
"BIO_SIZE_TO", "SIZE_UNIT", "BIO_SAMPLE_WEIGHT", "WEIGHT_UNIT",
"BIO_SAMPLE_WEIGHT_FROM", "BIO_SAMPLE_WEIGHT_TO", "WEIGHT_UNIT_1",
"SPECIES", "RESULT_LIFE_STAGE"}

The default value of \code{cols} is \code{"wq"}, which will return a data
frame with the following columns:

\code{"EMS_ID", "MONITORING_LOCATION", "LATITUDE", "LONGITUDE",
"LOCATION_TYPE", "COLLECTION_START", "LOCATION_PURPOSE", "PERMIT",
"SAMPLE_CLASS", "SAMPLE_STATE", "SAMPLE_DESCRIPTOR", "PARAMETER_CODE",
"PARAMETER", "ANALYTICAL_METHOD_CODE", "ANALYTICAL_METHOD", "RESULT_LETTER",
"RESULT", "UNIT", "METHOD_DETECTION_LIMIT", "QA_INDEX_CODE", "UPPER_DEPTH",
"LOWER_DEPTH"}
}
