% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{lt_lake_sites}
\alias{lt_lake_sites}
\title{Long-term lake monitoring sites}
\usage{
lt_lake_sites()
}
\value{
a character vector of ems ids
}
\description{
Get the EMS_IDs of all of the long-term lake monitoring sites.
}
