% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ems_data.R
\name{remove_data_cache}
\alias{remove_data_cache}
\title{Remove cached EMS data from your computer}
\usage{
remove_data_cache(which = c("all", "current", "historic"))
}
\arguments{
\item{which}{which data to remove? Either \code{"current"}, \code{"historic"},
or \code{"all"}.}
}
\description{
Remove cached EMS data from your computer
}

