% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{standardize_mdl_units}
\alias{standardize_mdl_units}
\title{Standardize MDL}
\usage{
standardize_mdl_units(data)
}
\arguments{
\item{data}{an ems data frame containing at least the
columns \verb{"UNIT", "METHOD_DETECTION_LIMIT", "MDL_UNIT"}}
}
\value{
data frame with MDLs standardized to UNITs (where possible)
}
\description{
There are many cases in EMS where the UNIT of the RESULT is
displayed differently than the METHOD_DETECTION_LIMIT.
This function uses the \code{units} package to convert the
\code{METHOD_DETECTION_LIMIT} values to the same unit as \code{RESULT}.
}
